requirements.txt
requests==2.31.0
python-dotenv==1.0.0
click==8.1.3
openai==0.27.8
